% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method1.R
\name{method1}
\alias{method1}
\title{Method 1 from the Ghosh and Crosby (2005) paper.}
\usage{
method1(D, threshold = 0, CAF = 0.5)
}
\arguments{
\item{D}{A vector of crossover differences}

\item{threshold}{If the largest gap is below this value, then do not form subgroups.}

\item{CAF}{A multiplicative factor that determines the critical value for forming subgroups.}
}
\value{
A list.
}
\description{
Method 1 from the Ghosh and Crosby (2005) paper.
}
